+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Changes
 
 - Reduces the bottom margin on labels and legends for form components from "16px" to "8px", or "s" to "sx" in Chakra-theme variables.
+- Adds "(Required)" text to the placeholder in the `SearchBar` component when `isRequired` is true.
 
 ## 0.25.5 (December 9, 2021)